moving library to lib and creating some unittests
