--- .travis.yml
@@ -1,27 +1,0 @@
-language: python
-python:
-    - "2.7"
-
-# Cache PlatformIO packages using Travis CI container-based infrastructure
-sudo: false
-cache:
-    directories:
-        - "~/.platformio"
-
-env:
-    - SCRIPT=platformioSingle EXAMPLE_NAME=ChannelStatistics EXAMPLE_FOLDER=/ BOARD=d1_mini
-    - SCRIPT=platformioSingle EXAMPLE_NAME=ChannelStatistics EXAMPLE_FOLDER=/ BOARD=tinypico
-
-install:
-    - pip install -U platformio
-    #
-    # Libraries from PlatformIO Library Registry:
-    #
-    # http://platformio.org/lib/show/64/ArduinoJson
-    - platformio lib -g install 64
-    # http://platformio.org/lib/show/567/WifiManager
-    - platformio lib -g install 567
-    # Double Reset Detector (not yet on PIO)
-    - platformio lib -g install https://github.com/datacute/DoubleResetDetector.git
-
-script: scripts/travis/$SCRIPT.sh
