+++ platformio.ini
@@ -11,13 +11,6 @@
 [common]
 lib_deps_external = ArduinoJson
 
-[env:d1_mini]
-platform = espressif8266
-board = d1_mini
-framework = arduino
-lib_deps = ${common.lib_deps_external}
-monitor_speed = 115200
-
 [env:esp32_devkit]
 platform = espressif32
 board = esp32doit-devkit-v1
