--- scripts/travis/platformioSingle.sh
@@ -1,3 +1,0 @@
-#!/bin/sh -eux
-
-platformio ci $PWD/examples/$EXAMPLE_NAME/$EXAMPLE_NAME.ino -l '.' -b $BOARD
